Move email logic again to backend
