+++ mobile/screens/Onboarding/HandlerInformationScreen.tsx
@@ -73,25 +73,6 @@ export default function HandlerInformationScreen(props: any) {
           default: "Failed to Update Handler Information",
         },
       });
-      const emailData = {
-        email: (user as User).email,
-        firstName: firstName,
-        lastName: lastName,
-        address: address,
-      }
-      await ErrorWrapper({
-        functionToExecute: sendEmail,
-        errorHandler: setError,
-        parameters: [
-          "gt.engineering@hack4impact.org",
-          EmailSubject.ACCOUNT_CREATED,
-          EmailTemplate.ACCOUNT_CREATED,
-          emailData
-        ],
-        customErrors: {
-          default: "Failed to Send Email",
-        },
-      })
       return user;
     } catch (e) {
       endOfExecutionHandler(e as Error);
